requirements.txt
Pillow>=10.0.0
rembg>=2.0.50
onnxruntime>=1.10.0
opencv-python>=4.5.0
numpy>=1.21.0
scikit-image>=0.19.0
svgwrite>=1.4.0
Flask>=2.0.0
Werkzeug>=2.0.0